third_party/blink/web_tests/external/wpt/html/dom/reflection-embedded-expected.txt

This is a testharness.js-based test.
Found 106 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] img.referrerPolicy: setAttribute() to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] img.referrerPolicy: setAttribute() to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] img.referrerPolicy: setAttribute() to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] img.referrerPolicy: setAttribute() to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] img.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] img.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] img.referrerPolicy: IDL set to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] img.referrerPolicy: IDL set to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] img.referrerPolicy: IDL set to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] img.referrerPolicy: IDL set to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] img.referrerPolicy: IDL set to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] img.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] iframe.allowUserMedia: typeof IDL attribute
  assert_equals: expected "boolean" but got "undefined"
[FAIL] iframe.allowUserMedia: IDL get with DOM attribute unset
  assert_equals: expected (boolean) false but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to ""
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to " foo "
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to undefined
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to null
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to 7
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to 1.5
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to "5%"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to "+100"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to ".5"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to true
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to false
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to object "[object Object]"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to NaN
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to Infinity
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to -Infinity
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to "\\0"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to object "test-toString"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to object "test-valueOf"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: setAttribute() to "allowUserMedia"
  assert_equals: IDL get expected (boolean) true but got (undefined) undefined
[FAIL] iframe.allowUserMedia: IDL set to ""
  assert_equals: hasAttribute() expected false but got true
[FAIL] iframe.allowUserMedia: IDL set to " foo "
  assert_equals: IDL get expected (boolean) true but got (string) " foo "
[FAIL] iframe.allowUserMedia: IDL set to undefined
  assert_equals: hasAttribute() expected false but got true
[FAIL] iframe.allowUserMedia: IDL set to null
  assert_equals: hasAttribute() expected false but got true
[FAIL] iframe.allowUserMedia: IDL set to 7
  assert_equals: IDL get expected (boolean) true but got (number) 7
[FAIL] iframe.allowUserMedia: IDL set to 1.5
  assert_equals: IDL get expected (boolean) true but got (number) 1.5
[FAIL] iframe.allowUserMedia: IDL set to "5%"
  assert_equals: IDL get expected (boolean) true but got (string) "5%"
[FAIL] iframe.allowUserMedia: IDL set to "+100"
  assert_equals: IDL get expected (boolean) true but got (string) "+100"
[FAIL] iframe.allowUserMedia: IDL set to ".5"
  assert_equals: IDL get expected (boolean) true but got (string) ".5"
[FAIL] iframe.allowUserMedia: IDL set to false
  assert_equals: hasAttribute() expected false but got true
[FAIL] iframe.allowUserMedia: IDL set to object "[object Object]"
  assert_equals: IDL get expected (boolean) true but got (object) object "[object Object]"
[FAIL] iframe.allowUserMedia: IDL set to NaN
  assert_equals: hasAttribute() expected false but got true
[FAIL] iframe.allowUserMedia: IDL set to Infinity
  assert_equals: IDL get expected (boolean) true but got (number) Infinity
[FAIL] iframe.allowUserMedia: IDL set to -Infinity
  assert_equals: IDL get expected (boolean) true but got (number) -Infinity
[FAIL] iframe.allowUserMedia: IDL set to "\\0"
  assert_equals: IDL get expected (boolean) true but got (string) "\\0"
[FAIL] iframe.allowUserMedia: IDL set to object "test-toString"
  assert_equals: IDL get expected (boolean) true but got (object) object "test-toString"
[FAIL] iframe.allowUserMedia: IDL set to object "test-valueOf"
  assert_equals: IDL get expected (boolean) true but got (object) object "test-valueOf"
[FAIL] iframe.referrerPolicy: setAttribute() to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] iframe.referrerPolicy: setAttribute() to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] iframe.referrerPolicy: setAttribute() to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] iframe.referrerPolicy: setAttribute() to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] iframe.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] iframe.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] iframe.referrerPolicy: IDL set to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] iframe.referrerPolicy: IDL set to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] iframe.referrerPolicy: IDL set to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] iframe.referrerPolicy: IDL set to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] iframe.referrerPolicy: IDL set to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] iframe.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] iframe.delegateStickyUserActivation: typeof IDL attribute
  assert_equals: expected "string" but got "undefined"
[FAIL] iframe.delegateStickyUserActivation: setAttribute() to "vibration"
  assert_equals: IDL get expected (string) "vibration" but got (undefined) undefined
[FAIL] iframe.delegateStickyUserActivation: setAttribute() to "VIBRATION"
  assert_equals: IDL get expected (string) "vibration" but got (undefined) undefined
[FAIL] iframe.delegateStickyUserActivation: setAttribute() to "media"
  assert_equals: IDL get expected (string) "media" but got (undefined) undefined
[FAIL] iframe.delegateStickyUserActivation: setAttribute() to "MEDIA"
  assert_equals: IDL get expected (string) "media" but got (undefined) undefined
[FAIL] iframe.delegateStickyUserActivation: IDL set to ""
  assert_equals: getAttribute() expected "" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo "
  assert_equals: getAttribute() expected " \\0\\x01\\x02\\x03\\x04\\x05\\x06\\x07 \\b\\t\\n\\v\\f\\r\\x0e\\x0f \\x10\\x11\\x12\\x13\\x14\\x15\\x16\\x17 \\x18\\x19\\x1a\\x1b\\x1c\\x1d\\x1e\\x1f  foo " but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to undefined
  assert_equals: getAttribute() expected "undefined" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to 7
  assert_equals: getAttribute() expected "7" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to 1.5
  assert_equals: getAttribute() expected "1.5" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "5%"
  assert_equals: getAttribute() expected "5%" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "+100"
  assert_equals: getAttribute() expected "+100" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to ".5"
  assert_equals: getAttribute() expected ".5" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to true
  assert_equals: getAttribute() expected "true" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to false
  assert_equals: getAttribute() expected "false" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to object "[object Object]"
  assert_equals: getAttribute() expected "[object Object]" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to NaN
  assert_equals: getAttribute() expected "NaN" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to Infinity
  assert_equals: getAttribute() expected "Infinity" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to -Infinity
  assert_equals: getAttribute() expected "-Infinity" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "\\0"
  assert_equals: getAttribute() expected "\\0" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to object "test-toString"
  assert_equals: getAttribute() expected "test-toString" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to object "test-valueOf"
  assert_equals: getAttribute() expected "test-valueOf" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "vibration"
  assert_equals: getAttribute() expected "vibration" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "xvibration"
  assert_equals: getAttribute() expected "xvibration" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "vibration\\0"
  assert_equals: getAttribute() expected "vibration\\0" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "ibration"
  assert_equals: getAttribute() expected "ibration" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "VIBRATION"
  assert_equals: getAttribute() expected "VIBRATION" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "media"
  assert_equals: getAttribute() expected "media" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "xmedia"
  assert_equals: getAttribute() expected "xmedia" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "media\\0"
  assert_equals: getAttribute() expected "media\\0" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "edia"
  assert_equals: getAttribute() expected "edia" but got "MEDIA"
[FAIL] iframe.delegateStickyUserActivation: IDL set to "MEDIA"
  assert_equals: IDL get expected "media" but got "MEDIA"
[FAIL] area.referrerPolicy: setAttribute() to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] area.referrerPolicy: setAttribute() to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] area.referrerPolicy: setAttribute() to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] area.referrerPolicy: setAttribute() to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] area.referrerPolicy: setAttribute() to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] area.referrerPolicy: setAttribute() to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] area.referrerPolicy: IDL set to "same-origin"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] area.referrerPolicy: IDL set to "SAME-ORIGIN"
  assert_equals: IDL get expected "same-origin" but got ""
[FAIL] area.referrerPolicy: IDL set to "strict-origin"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] area.referrerPolicy: IDL set to "STRICT-ORIGIN"
  assert_equals: IDL get expected "strict-origin" but got ""
[FAIL] area.referrerPolicy: IDL set to "strict-origin-when-cross-origin"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
[FAIL] area.referrerPolicy: IDL set to "STRICT-ORIGIN-WHEN-CROSS-ORIGIN"
  assert_equals: IDL get expected "strict-origin-when-cross-origin" but got ""
Harness: the test ran to completion.
